restructure storage config
